==> 1_3_Converting_xyz_to_mol.ipynb <==
from rdkit import Chem
import pandas as pd
import os
from rdkit.Geometry import Point3D

data = pd.read_csv('./Caco-2 Permeability Dataset/Filtered Data/3D_optimizable_dataset.csv')

for idx in range(len(data)):
    atoms = []
    xyz_optimized_coords = []
    xyz_file = f'./3D Conformers/xtb/xyz/xtbopt_{idx}.xyz'

    if os.path.exists(xyz_file):
        with open(xyz_file, 'r') as f:
            for line_number,line in enumerate(f):
                if line_number == 0:
                    num_atoms = int(line)
                elif line_number == 1:
                    comment = line
                else:
                    atomic_symbol, x, y, z = line.split()
                    atoms.append(atomic_symbol)
                    xyz_optimized_coords.append([float(x),float(y),float(z)])
        
        m = Chem.MolFromMolFile(f'./3D Conformers/rdkit/mol/mol_{idx}.mol', removeHs=False)
        conf = m.GetConformer()
        for i in range(m.GetNumAtoms()):
           x,y,z = xyz_optimized_coords[i]
           conf.SetAtomPosition(i,Point3D(x,y,z))

        Chem.MolToMolFile(m, f'./3D Conformers/xtb/mol/xtbopt_{idx}.mol')
